WIP: extracting random import functions into a delegate.
