+++ ehri-io/src/main/java/eu/ehri/project/importers/base/AbstractImporter.java
@@ -19,8 +19,6 @@
 
 package eu.ehri.project.importers.base;
 
-import com.google.common.base.Charsets;
-import com.google.common.base.Joiner;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
 import com.tinkerpop.frames.FramedGraph;
@@ -29,20 +27,12 @@ import eu.ehri.project.core.GraphManagerFactory;
 import eu.ehri.project.exceptions.ValidationError;
 import eu.ehri.project.importers.ImportCallback;
 import eu.ehri.project.importers.ImportLog;
-import eu.ehri.project.importers.properties.NodeProperties;
-import eu.ehri.project.models.EntityClass;
 import eu.ehri.project.models.base.Accessible;
 import eu.ehri.project.models.base.Actioner;
 import eu.ehri.project.models.base.PermissionScope;
 import eu.ehri.project.persistence.BundleManager;
 import eu.ehri.project.persistence.Mutation;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.util.List;
 import java.util.Map;
 
@@ -56,10 +46,6 @@ import java.util.Map;
 public abstract class AbstractImporter<T> {
 
     public static final String OBJECT_IDENTIFIER = "objectIdentifier";
-    private static final String NODE_PROPERTIES = "allowedNodeProperties.csv";
-
-    private static final Logger logger = LoggerFactory.getLogger(AbstractImporter.class);
-    private static final Joiner stringJoiner = Joiner.on("\n\n").skipNulls();
 
     protected final PermissionScope permissionScope;
     protected final Actioner actioner;
@@ -67,8 +53,6 @@ public abstract class AbstractImporter<T> {
     protected final GraphManager manager;
     protected final ImportLog log;
     private final List<ImportCallback> callbacks = Lists.newArrayList();
-
-    private NodeProperties pc;
 
     /**
      * Call all registered ImportCallbacks for the given mutation.
@@ -153,49 +137,7 @@ public abstract class AbstractImporter<T> {
     public abstract Iterable<Map<String, Object>> extractDates(T data);
 
 
-    /**
-     * only properties that have the multivalued-status can actually be multivalued. all other properties will be
-     * flattened by this method.
-     *
-     * @param key    a property key
-     * @param value  a property value
-     * @param entity the EntityClass with which this frameMap must comply
-     */
-    protected Object flattenNonMultivaluedProperties(String key, Object value, EntityClass entity) {
-        if (pc == null) {
-            pc = loadNodeProperties();
-        }
-        if (value instanceof List
-                && !(pc.hasProperty(entity.getName(), key) && pc.isMultivaluedProperty(entity.getName(), key))) {
-            logger.trace("Flattening array property value: {}: {}", key, value);
-            return stringJoiner.join((List) value);
-        } else {
-            return value;
-        }
-    }
-
     public abstract Iterable<Map<String, Object>> extractMaintenanceEvent(T itemData);
 
     public abstract Map<String, Object> getMaintenanceEvent(T event);
-
-    // Helpers
-
-    private NodeProperties loadNodeProperties() {
-        try (InputStream fis = getClass().getClassLoader().getResourceAsStream(NODE_PROPERTIES);
-             BufferedReader br = new BufferedReader(new InputStreamReader(fis, Charsets.UTF_8))) {
-            NodeProperties nodeProperties = new NodeProperties();
-            String headers = br.readLine();
-            nodeProperties.setTitles(headers);
-
-            String line;
-            while ((line = br.readLine()) != null) {
-                nodeProperties.addRow(line);
-            }
-            return nodeProperties;
-        } catch (IOException ex) {
-            throw new RuntimeException(ex);
-        } catch (NullPointerException npe) {
-            throw new RuntimeException("Missing or empty properties file: " + NODE_PROPERTIES);
-        }
-    }
 }
